Reorganize internal classes
