+++ intellij-plugin-structure/structure-ide/src/main/java/com/jetbrains/plugin/structure/ide/layout/LayoutComponentLoader.kt
@@ -1,15 +1,14 @@
 package com.jetbrains.plugin.structure.ide.layout
 
-import com.jetbrains.plugin.structure.ide.ProductInfoBasedIdeManager
 import com.jetbrains.plugin.structure.intellij.resources.ResourceResolver
 import com.jetbrains.plugin.structure.intellij.version.IdeVersion
 import java.nio.file.Path
 
-fun interface LayoutComponentLoader {
+internal fun interface LayoutComponentLoader {
   fun load(
     pluginArtifactPath: Path,
     descriptorName: String,
     resourceResolver: ResourceResolver,
     ideVersion: IdeVersion
-  ): ProductInfoBasedIdeManager.PluginWithArtifactPathResult
+  ): PluginWithArtifactPathResult
 }
